fix(comment): resolve comment issues.
