--- src/input-button/index.ts
@@ -1,1 +1,0 @@
-export { default, default as InputButton } from './InputButton';
