--- src/input-button/style/index.ts
@@ -1,1 +1,0 @@
-import './index.less';
