+++ src/input/index.ts
@@ -1,4 +1,5 @@
 import Input from './Input';
+import InputButton from './InputButton';
 import InputNumber from './InputNumber';
 import Password from './Password';
 import TextArea from './TextArea';
@@ -6,6 +7,7 @@ import TextArea from './TextArea';
 Input.InputNumber = InputNumber;
 Input.Password = Password;
 Input.TextArea = TextArea;
+Input.Button = InputButton;
 
 export * from './interface';
