+++ src/input/style/index.ts
@@ -1,1 +1,2 @@
 import './index.less';
+import './InputButton.less';
