+++ src/stylesheet/index.less
@@ -19,7 +19,6 @@
 @margin-xss: 4px; // more small
 
 @border-color-base: #a3adc8;
-@border-color-gray2: #dfe4ee;
 @border-color-split: hsv(0, 0, 94%); // split border inside a component
 @border-color-inverse: #fff;
 @border-width-base: 1px; // width of the border for a component
